Lines on non-visible data fix
In order to draw the lines on a certain point, we're using a 'beginDrawDateIndex' and an 'endDrawDateIndex'. This index references a set of values in our data-set (array). When looping through the data set, we take the date value of each data-record and compare it to the date set in Google Sheets. Given that the array is sorted, the index of the first matching item is set as the beginDrawDateIndex. We continue through the loop and set the endDrawDateIndex for every next match. This means that once the date no longer matches, the endDrawDateIndex will no longer be set, leaving us with the index of the last matching data value.

With the above in mind: We'd initially programmed the drawing of lines with a fallback, that if a date was not set, a default value would be used. The default value has now been removed, which should remove some -if not all- of the excess lines in the chart.

Additionally, a startDateDaysLoaded variable has been introduced as an extra check. This value is used to compare against the date on the data record, and will refrain from drawing a line on this record if the record is outside of the view. The value is set by taking the BaseDateTimeIn of the last data record, and subtracting the DaysToLoadInChart.